Rename editor.rs -> mod.rs
