+++ src/views/editor/actions.rs
@@ -9,8 +9,7 @@ use floem_editor_core::{
 
 use super::{
     command::{Command, CommandExecuted},
-    editor::Editor,
-    movement,
+    movement, Editor,
 };
 
 pub fn handle_command_default(
